+++ docs/news.rst
@@ -15,6 +15,7 @@ Psycopg 3.3.0 (unreleased)
 
 .. rubric:: New top-level features
 
+- Add :ref:`template strings queries <template-strings>` (:ticket:`#1054`).
 - Cursors are now iterators, not only iterables. This means you can call
   ``next(cur)`` to fetch the next row (:ticket:`#1064`).
 - Add `Cursor.results()` to iterate over the result sets of the queries